# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "T-Racks OpenApi Documentation",
        "description": "T-Racks Swagger OpenApi description",
        "contact": {
            "email": "vincent@vinniict.nl"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://t-racks.localhost/api/v1",
            "description": "T-Racks Server"
        }
    ],
    "paths": {
        "/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get list of locations",
                "description": "Returns list of all locations",
                "operationId": "getLocationsList",
                "responses": {
                    "200": {
                        "description": "Succesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Store new location",
                "description": "Returns location data",
                "operationId": "storeLocation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get location information",
                "description": "Returns location data",
                "operationId": "getLocationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update existing location",
                "description": "Returns updated location data",
                "operationId": "updateLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rows": {
            "get": {
                "tags": [
                    "Rows"
                ],
                "summary": "Get list of rows",
                "description": "Returns list of all rows",
                "operationId": "getRowsList",
                "responses": {
                    "200": {
                        "description": "Succesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RowResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rows"
                ],
                "summary": "Store new row",
                "description": "Returns row data",
                "operationId": "storeRow",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRowRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Row"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rows/{id}": {
            "get": {
                "tags": [
                    "Rows"
                ],
                "summary": "Get row information",
                "description": "Returns row data",
                "operationId": "getrowById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Row id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Row"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Rows"
                ],
                "summary": "Update existing rows",
                "description": "Returns updated rows data",
                "operationId": "updateRow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Row id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRowRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Row"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error404": {
                "title": "404",
                "description": "404 model",
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "The error message",
                        "type": "string",
                        "example": "Not Found!"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "404"
                }
            },
            "Error422": {
                "title": "422",
                "description": "422 model",
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "The error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "422"
                }
            },
            "Location": {
                "title": "Location",
                "description": "Location model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "The name of the location",
                        "type": "string",
                        "example": "Location 1"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Location"
                }
            },
            "Row": {
                "title": "Row",
                "description": "Row model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "The name of the row",
                        "type": "string",
                        "example": "Row 1"
                    },
                    "location_id": {
                        "title": "Location ID",
                        "description": "The ID of the location",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Row"
                }
            },
            "LocationResource": {
                "title": "Location Resource",
                "description": "Location resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Location"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Location Resource"
                }
            },
            "RowResource": {
                "title": "row Resource",
                "description": "Row resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Row"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Row Resource"
                }
            },
            "StoreLocationRequest": {
                "title": "Store Location Request",
                "description": "Store Locations request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new location",
                        "type": "string",
                        "example": "Location 1"
                    }
                },
                "type": "object"
            },
            "StoreRowRequest": {
                "title": "Store Row Request",
                "description": "Store Row request body data",
                "required": [
                    "location_id"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new row",
                        "type": "string",
                        "example": "Row 1"
                    },
                    "location_id": {
                        "title": "location_id",
                        "description": "ID of the location",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "UpdateLocationRequest": {
                "title": "Update Location Request",
                "description": "Update Locations request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "New name of the location",
                        "type": "string",
                        "example": "Location 1"
                    }
                },
                "type": "object"
            },
            "UpdateRowRequest": {
                "title": "Update Row Request",
                "description": "Update Row request body data",
                "required": [
                    "location_id"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "New name of the row",
                        "type": "string",
                        "example": "Row 1"
                    },
                    "location_id": {
                        "title": "location_id",
                        "description": "New location_id of the row",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Locations",
            "description": "API Endpoints of Locations"
        },
        {
            "name": "Rows",
            "description": "API Endpoints of Rows"
        }
    ],
    "security": [
        []
    ]
}